pyproject.toml
[project]
name = "csun-analytics"
version = "0.1.0"
description = "Analytics toolkit for the CSUN Assistive Technology Conference"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pandas>=2.2",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "anthropic>=0.40",
    "tqdm>=4.66",
    "rich>=13.7",
    "plotly>=5.20",
    "networkx>=3.2",
    "kaleido>=0.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "ipython>=8.0",
    "jupyter>=1.0",
]
browser = [
    "playwright>=1.40",
]
dashboard = [
    "dash>=2.17",
    "dash-bootstrap-components>=1.6",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]
all = [
    "csun-analytics[dev,browser,dashboard,docs]",
]

[project.scripts]
csun = "csun_analytics.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/csun_analytics"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
